# docs/swagger.yaml
basePath: /
definitions:
...
  handler.RespBody:
    properties:
      code:
        description: http code
        type: integer
      data:
        description: response data
      msg:
        description: response message
        type: string
      reason:
        description: reason key
        type: string
    type: object
...
  schema.AcceptAnswerReq:
    properties:
      answer_id:
        type: string
      question_id:
        maxLength: 30
        type: string
    required:
    - question_id
    type: object
...
  schema.AdminUpdateAnswerStatusReq:
    properties:
      answer_id:
        type: string
      status:
        enum:
        - available
        - deleted
        type: string
    required:
    - answer_id
    - status
    type: object
...
  schema.AnswerAddReq:
    properties:
      captcha_code:
        type: string
      captcha_id:
        type: string
      content:
        maxLength: 65535
        minLength: 6
        type: string
      question_id:
        type: string
    required:
    - content
    type: object
  schema.AnswerInfo:
    properties:
      accepted:
        type: integer
      collected:
        type: boolean
      content:
        type: string
      create_time:
        type: integer
      html:
        type: string
      id:
        type: string
      member_actions:
        description: MemberActions
        items:
          $ref: '#/definitions/schema.PermissionMemberAction'
        type: array
      question_id:
        type: string
      question_info:
        $ref: '#/definitions/schema.QuestionInfoResp'
      status:
        type: integer
      update_time:
        type: integer
      update_user_info:
        $ref: '#/definitions/schema.UserBasicInfo'
      user_info:
        $ref: '#/definitions/schema.UserBasicInfo'
      vote_count:
        type: integer
      vote_status:
        type: string
    type: object
  schema.AnswerUpdateReq:
    properties:
      captcha_code:
        type: string
      captcha_id:
        type: string
      content:
        maxLength: 65535
        minLength: 6
        type: string
      edit_summary:
        type: string
      id:
        type: string
      title:
        type: string
    required:
    - content
    type: object
...
  schema.GetAnswerInfoResp:
    properties:
      info:
        $ref: '#/definitions/schema.AnswerInfo'
      question:
        $ref: '#/definitions/schema.QuestionInfoResp'
    type: object
...
  schema.Operation:
    properties:
      description:
        type: string
      level:
        $ref: '#/definitions/schema.OperationLevel'
      msg:
        type: string
      time:
        type: integer
      type:
        type: string
    type: object
  schema.OperationLevel:
    enum:
    - info
    - danger
    - warning
    - secondary
    type: string
    x-enum-varnames:
    - OperationLevelInfo
    - OperationLevelDanger
    - OperationLevelWarning
    - OperationLevelSecondary
...
  schema.PermissionMemberAction:
    properties:
      action:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
...
  schema.QuestionInfoResp:
    properties:
      accepted_answer_id:
        type: string
      answer_count:
        type: integer
      answered:
        type: boolean
      collected:
        type: boolean
      collection_count:
        type: integer
      content:
        type: string
      create_time:
        type: integer
      description:
        type: string
      edit_time:
        type: integer
      extends_actions:
        items:
          $ref: '#/definitions/schema.PermissionMemberAction'
        type: array
      first_answer_id:
        type: string
      follow_count:
        type: integer
      html:
        type: string
      id:
        type: string
      is_followed:
        type: boolean
      last_answer_id:
        type: string
      last_answered_user_info:
        $ref: '#/definitions/schema.UserBasicInfo'
      member_actions:
        description: MemberActions
        items:
          $ref: '#/definitions/schema.PermissionMemberAction'
        type: array
      operation:
        $ref: '#/definitions/schema.Operation'
      pin:
        type: integer
      section_id:
        type: integer
      show:
        type: integer
      status:
        type: integer
      tags:
        items:
          $ref: '#/definitions/schema.TagResp'
        type: array
      title:
        type: string
      unique_view_count:
        type: integer
      update_time:
        type: integer
      update_user_info:
        $ref: '#/definitions/schema.UserBasicInfo'
      url_title:
        type: string
      user_info:
        $ref: '#/definitions/schema.UserBasicInfo'
      view_count:
        type: integer
      vote_count:
        type: integer
      vote_status:
        type: string
    type: object
...
  schema.RecoverAnswerReq:
    properties:
      answer_id:
        type: string
    required:
    - answer_id
    type: object
...
  schema.RemoveAnswerReq:
    properties:
      captcha_code:
        type: string
      captcha_id:
        type: string
      id:
        type: string
    required:
    - id
    type: object
...
  schema.TagResp:
    properties:
      display_name:
        type: string
      main_tag_slug_name:
        description: if main tag slug name is not empty, this tag is synonymous with
          the main tag
        type: string
      recommend:
        type: boolean
      reserved:
        type: boolean
      slug_name:
        type: string
    type: object
...
  schema.UserBasicInfo:
    properties:
      avatar:
        type: string
      display_name:
        type: string
      id:
        type: string
      language:
        type: string
      location:
        type: string
      rank:
        type: integer
      status:
        type: string
      suspended_until:
        type: integer
      username:
        type: string
      website:
        type: string
    type: object
...
info:
  contact: {}
  description: Apache Answer API
  title: Apache Answer
paths:
...
  /answer/admin/api/answer/status:
    put:
      consumes:
      - application/json
      description: update answer status
      parameters:
      - description: AdminUpdateAnswerStatusReq
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.AdminUpdateAnswerStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update answer status
      tags:
      - admin
...
  /answer/api/v1/answer:
    delete:
      consumes:
      - application/json
      description: delete answer
      parameters:
      - description: answer
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.RemoveAnswerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: delete answer
      tags:
      - Answer
    post:
      consumes:
      - application/json
      description: add answer
      parameters:
      - description: add answer request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.AnswerAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: Add Answer
      tags:
      - Answer
    put:
      consumes:
      - application/json
      description: Update Answer
      parameters:
      - description: AnswerUpdateReq
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.AnswerUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: Update Answer
      tags:
      - Answer
  /answer/api/v1/answer/acceptance:
    post:
      consumes:
      - application/json
      description: Accept Answer
      parameters:
      - description: AcceptAnswerReq
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.AcceptAnswerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: Accept Answer
      tags:
      - Answer
  /answer/api/v1/answer/info:
    get:
      consumes:
      - application/json
      description: Get Answer Detail
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.GetAnswerInfoResp'
              type: object
      summary: Get Answer Detail
      tags:
      - Answer
  /answer/api/v1/answer/page:
    get:
      consumes:
      - application/json
      description: AnswerList <br> <b>order</b> (default or updated)
      parameters:
      - description: question_id
        in: query
        name: question_id
        required: true
        type: string
      - description: order
        in: query
        name: order
        required: true
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: page_size
        in: query
        name: page_size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: AnswerList
      tags:
      - Answer
  /answer/api/v1/answer/recover:
    post:
      consumes:
      - application/json
      description: recover the deleted answer
      parameters:
      - description: answer
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.RecoverAnswerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: recover answer
      tags:
      - Answer
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"